feat: Add medicine list management components including MedicineCard, MedicineCounter, DeleteButton, EmptyState, and MedicationInfo
