+++ package.json
@@ -21,7 +21,6 @@
     "motion": "^12.4.7",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
-    "react-helmet": "^6.1.0",
     "react-icons": "^5.4.0",
     "react-router-dom": "^7.3.0",
     "tailwindcss": "^4.0.7"
